requirements.txt
Flask
requests
pymongo
mysql-connector-python
google-generativeai
gunicorn
PyJWT
sentence-transformers
torch